argon2: First pass for supporting Argon2id as well
