+++ src/hazardous/kdf/mod.rs
@@ -33,5 +33,14 @@ pub mod scrypt;
 
 #[cfg(any(feature = "safe_api", feature = "alloc"))]
 #[cfg_attr(docsrs, doc(cfg(any(feature = "safe_api", feature = "alloc"))))]
+mod argon2;
+
+#[cfg(any(feature = "safe_api", feature = "alloc"))]
+#[cfg_attr(docsrs, doc(cfg(any(feature = "safe_api", feature = "alloc"))))]
 /// Argon2i password hashing function as described in the [P-H-C specification](https://github.com/P-H-C/phc-winner-argon2/blob/master/argon2-specs.pdf).
-pub mod argon2i;
+pub use argon2::argon2i;
+
+#[cfg(any(feature = "safe_api", feature = "alloc"))]
+#[cfg_attr(docsrs, doc(cfg(any(feature = "safe_api", feature = "alloc"))))]
+/// Argon2id password hashing function as described in the [P-H-C specification](https://github.com/P-H-C/phc-winner-argon2/blob/master/argon2-specs.pdf).
+pub use argon2::argon2id;
