+++ tests/kdf/other_argon2i.rs
@@ -21,6 +21,7 @@ fn test_case_0() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -34,6 +35,7 @@ fn test_case_0() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -60,6 +62,7 @@ fn test_case_1() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -73,6 +76,7 @@ fn test_case_1() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -99,6 +103,7 @@ fn test_case_2() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -112,6 +117,7 @@ fn test_case_2() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -138,6 +144,7 @@ fn test_case_3() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -151,6 +158,7 @@ fn test_case_3() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -177,6 +185,7 @@ fn test_case_4() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -190,6 +199,7 @@ fn test_case_4() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -216,6 +226,7 @@ fn test_case_5() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -229,6 +240,7 @@ fn test_case_5() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -255,6 +267,7 @@ fn test_case_6() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -268,6 +281,7 @@ fn test_case_6() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -294,6 +308,7 @@ fn test_case_7() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -307,6 +322,7 @@ fn test_case_7() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -332,6 +348,7 @@ fn test_case_8() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -345,6 +362,7 @@ fn test_case_8() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -370,6 +388,7 @@ fn test_case_9() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -383,6 +402,7 @@ fn test_case_9() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -409,6 +429,7 @@ fn test_case_10() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -422,6 +443,7 @@ fn test_case_10() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -448,6 +470,7 @@ fn test_case_11() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -461,6 +484,7 @@ fn test_case_11() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -487,6 +511,7 @@ fn test_case_12() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -500,6 +525,7 @@ fn test_case_12() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -526,6 +552,7 @@ fn test_case_13() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -539,6 +566,7 @@ fn test_case_13() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
@@ -565,6 +593,7 @@ fn test_case_14() {
         &salt,
         passes,
         mem,
+        1,
         Some(secret_value),
         Some(associated_data),
         &mut actual_hash,
@@ -578,6 +607,7 @@ fn test_case_14() {
             &salt,
             passes,
             mem,
+            1,
             Some(secret_value),
             Some(associated_data),
             &mut actual_hash
